third_party/blink/web_tests/flag-specific/highdpi/external/wpt/css/mediaqueries/mq-invalid-media-type-layer-002-expected.txt

This is a testharness.js-based test.
FAIL 'layer' used as media types is a syntax error assert_equals: expected "not all" but got "not layer"
Harness: the test ran to completion.
